fix: clamp fullscreen dashboard scroll to its own content
clampPreviewScroll computed maxScroll from the right-column preview
(narrow width, plus previewEventsContent / metrics footer), which has
nothing to do with the fullscreen dashboard. So in dashboard mode the
scroll offset could run past the last rendered line into blank space.

Add clampDashboardScroll, which bounds previewScroll against the
dashboard's own content (single-col line count, or the taller of the two
columns) and viewport, and route clampPreviewScroll to it whenever
fullscreenDashboard is active so every scroll entry point (arrows,
ctrl+d/u, ctrl+f/b, G, J/K) is bounded consistently.

Extract the dashboard rendering + scroll helpers into view_dashboard.go
to keep view.go under the 800-line cap.
